+++ pixeltable/share/packager.py
